Langchain/requirements.txt
langchain
python-dotenv
langchain-community
pypdf
bs4
arxiv
PyMuPDF 
wikipedia
langchain_openai
chromadb
sentence_transformers
langchain_huggingface
tf-keras
faiss-cpu
langchain_chroma
beautifulsoup4
langchain_google_genai
langchain_core
streamlit
langchain_groq
fastapi
uvicorn
sse_starlette
langserve
